+++ CHANGELOG.md
@@ -22,6 +22,8 @@ All notable changes to this project will be documented in this file.
 ## [0.8] - 22.05.2022
 
 ### Added
+- `CameraPane` component.
+- `HexagonView` component.
 - `HexagonGrid` with two coordinate systems. axial and offset coordinates.
 
 ## [0.7.3] - 31.08.2022